requirements.txt
boto3
Pillow
tkinter